requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-filter==23.1
paho-mqtt==1.6.1
python-decouple==3.8
celery==5.3.0
redis==5.0.0
psycopg2-binary==2.9.6
gunicorn==21.2.0
